+++ components/core/src/clp/WildcardExpression.hpp
@@ -1,7 +1,7 @@
 #ifndef CLP_WILDCARDEXPRESSION_HPP
 #define CLP_WILDCARDEXPRESSION_HPP
 
-#include <cstdint>
+#include <cstddef>
 #include <string>
 #include <vector>
 
@@ -19,26 +19,25 @@ class WildcardExpression {
 public:
     explicit WildcardExpression(std::string processed_search_string);
 
-    [[nodiscard]] auto
-    substr(uint32_t const begin_idx, uint32_t const length) const -> std::string {
+    [[nodiscard]] auto substr(size_t const begin_idx, size_t const length) const -> std::string {
         return m_processed_search_string.substr(begin_idx, length);
     }
 
-    [[nodiscard]] auto length() const -> uint32_t { return m_processed_search_string.size(); }
+    [[nodiscard]] auto length() const -> size_t { return m_processed_search_string.size(); }
 
-    [[nodiscard]] auto get_value_is_greedy_wildcard(uint32_t const idx) const -> bool {
+    [[nodiscard]] auto get_value_is_greedy_wildcard(size_t const idx) const -> bool {
         return m_is_greedy_wildcard[idx];
     }
 
-    [[nodiscard]] auto get_value_is_non_greedy_wildcard(uint32_t const idx) const -> bool {
+    [[nodiscard]] auto get_value_is_non_greedy_wildcard(size_t const idx) const -> bool {
         return m_is_non_greedy_wildcard[idx];
     }
 
-    [[nodiscard]] auto get_value_is_escape(uint32_t const idx) const -> bool {
+    [[nodiscard]] auto get_value_is_escape(size_t const idx) const -> bool {
         return m_is_escape[idx];
     }
 
-    [[nodiscard]] auto get_value(uint32_t const idx) const -> char {
+    [[nodiscard]] auto get_value(size_t const idx) const -> char {
         return m_processed_search_string[idx];
     }
 
@@ -64,8 +63,8 @@ public:
      */
     WildcardExpressionView(
             WildcardExpression const& wildcard_expression,
-            uint32_t begin_idx,
-            uint32_t end_idx
+            size_t begin_idx,
+            size_t end_idx
     );
 
     /**
@@ -95,21 +94,21 @@ public:
     [[nodiscard]] auto surrounded_by_delims_or_wildcards(log_surgeon::lexers::ByteLexer const& lexer
     ) const -> bool;
 
-    [[nodiscard]] auto length() const -> uint32_t { return m_end_idx - m_begin_idx; }
+    [[nodiscard]] auto length() const -> size_t { return m_end_idx - m_begin_idx; }
 
-    [[nodiscard]] auto get_value_is_greedy_wildcard(uint32_t const idx) const -> bool {
+    [[nodiscard]] auto get_value_is_greedy_wildcard(size_t const idx) const -> bool {
         return m_search_string_ptr->get_value_is_greedy_wildcard(m_begin_idx + idx);
     }
 
-    [[nodiscard]] auto get_value_is_non_greedy_wildcard(uint32_t const idx) const -> bool {
+    [[nodiscard]] auto get_value_is_non_greedy_wildcard(size_t const idx) const -> bool {
         return m_search_string_ptr->get_value_is_non_greedy_wildcard(m_begin_idx + idx);
     }
 
-    [[nodiscard]] auto get_value_is_escape(uint32_t const idx) const -> bool {
+    [[nodiscard]] auto get_value_is_escape(size_t const idx) const -> bool {
         return m_search_string_ptr->get_value_is_escape(m_begin_idx + idx);
     }
 
-    [[nodiscard]] auto get_value(uint32_t const idx) const -> char {
+    [[nodiscard]] auto get_value(size_t const idx) const -> char {
         return m_search_string_ptr->get_value(m_begin_idx + idx);
     }
 
@@ -119,8 +118,8 @@ public:
 
 private:
     WildcardExpression const* m_search_string_ptr;
-    uint32_t m_begin_idx;
-    uint32_t m_end_idx;
+    size_t m_begin_idx;
+    size_t m_end_idx;
 };
 }  // namespace clp
